requirements.txt
djangorestframework~=3.12.1
Django~=3.1.2